pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/politician_embeddings"]

[project]
name = "politician-embeddings"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "plotly>=6.6.0",
    "requests>=2.32.5",
    "streamlit>=1.19.0",
    "torch>=2.10.0",
    "umap-learn>=0.5.11",
]

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D", "ANN", "COM812"]  # COM812 conflicts with ruff formatter

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "ruff>=0.15.5",
    "ty>=0.0.21",
]
